Implement primitive error handling
The basic idea is to define a collection of specifications, which can
cover contractual obligations (i.e. argument types), arity obligations
(either a specific arity, min arity or max), and any custom
specifications (for example, not dividing by zero).

Each of the specifications provides a proc to test the argument list,
and then a function to generate the correct error message when the
specification is violated